+++ stars/ui.go
@@ -122,7 +122,7 @@ func (sp *STARSPane) DrawInfo(c *client.ControlClient, p platform.Platform, lg *
 					imgui.Text(name)
 
 					imgui.TableNextColumn()
-					imgui.Text(strings.Join(arr.ServedAirports(), ", "))
+					imgui.Text(strings.Join(arr.Airports, ", "))
 
 					imgui.TableNextColumn()
 					if arr.Description != "" {